統一Configツリー・trainerライブラリ化・モデルストアを導入
- mars_lite/config.py: DataConfig/FeatureConfig/EnvConfig/TrainConfig/
  ServeConfig/RunConfigを新設。docs/ARCHITECTURE.mdの実測既定値
  （gamma=0.5, lambda_turnover=0.04, dd_derisk_start=0.10等）を単一の
  正として集約し、JSON往復・モデルメタデータ保存に対応
- tests/test_config.py: 実測既定値をピン留めする憲法テストを追加。
  既定値の変更が意図せず紛れ込むとテストが落ちる
- mars_lite/learning/trainer.py: train_ppo/make_env_fns を
  scripts/train_portfolio.py から移動。server/training_manager.py が
  sys.path操作でスクリプトをライブラリとしてimportしていた問題を解消し、
  正規のモジュールimportに変更
- mars_lite/pipeline/phases.py: phase_p0/train/wf/pbt/regime と
  build_post_processor/build_env_kwargs/build_feature_set を移動。
  scripts/train_portfolio.py は引数パース+フェーズ選択のみの
  薄いCLI(約150行)に縮小
- mars_lite/serving/model_store.py: 削除したlearning/model_manager.pyの
  後継。save_bundle/load_metadata/list_models/promote/rollbackを提供し、
  training_manager.py の手書きJSON保存と/api/signal/latestの手書きJSON
  読み込みを共通化。phase_trainの単一モデル保存もmodel_store経由にし、
  これまでCLI学習モデルにメタデータsidecarが無くserve側から使えなかった
  ギャップを解消
- scripts/train_portfolio.py: --min-trade-deltaのhelp文字列で'%未'が
  argparseのフォーマット文字列として誤解釈されクラッシュしていた
  pre-existingバグを修正(--helpが常に落ちていた)

p0ベンチマーク(--phase p0 --timesteps 20000 --seed 0)はPhase 0時点との
比較でビット一致。pytest 137 passed / 5 skipped。server経由の学習
(TrainingManager)とphase train両方でmodel_store保存を確認し、
/api/signal/latestをTestClientで実行して200応答・後処理・ガードレール
まで一貫して動くことを確認した。
