+++ mars_lite/pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""学習フェーズ実行パイプライン（p0/train/wf/pbt/regime）"""
