+++ mars_lite/serving/__init__.py
@@ -1,0 +1,1 @@
+"""運用・モデル永続化パッケージ"""
